Add option to prune snap path when building wheelhouse
When running charm-tools from a snap, the default is to use Python
and associated tools from the snap when building a charm's
wheelhouse etc.

This is normally not what you would want as the Python version in
the snap does not necessarily match the Python version of the
target system.

The wheelhouse needs to be built with a Python version compatible
with the Python version of the target system on which the built
charm is to be run on.

Add `--use-python-from-snap` option where the default is 'False'.
This is a change of behavior and mandates a major version bump for
this tool.

Fixes #623